+++ src/AsyncRepositories/ICRUDAsyncRepositoryBase.cs
@@ -1,6 +1,6 @@
 ﻿using PowerUtils.BuildingBlocks.Domain;
 
-namespace PowerUtils.BuildingBlocks.Data.Repositories
+namespace PowerUtils.BuildingBlocks.Data.AsyncRepositories
 {
     public interface ICRUDAsyncRepositoryBase<TEntity, TId> :
         IAddAsyncRepositoryBase<TEntity, TId>,
