+++ src/AsyncRepositories/ICountAsyncRepositoryBase.cs
@@ -2,13 +2,13 @@
 using System.Threading.Tasks;
 using PowerUtils.BuildingBlocks.Domain;
 
-namespace PowerUtils.BuildingBlocks.Data.Repositories
+namespace PowerUtils.BuildingBlocks.Data.AsyncRepositories
 {
-    public interface IDeleteAsyncRepositoryBase<TEntity, TId> :
+    public interface ICountAsyncRepositoryBase<TEntity, TId> :
         IEntityAsyncRepositoryBase<TEntity, TId>
 
         where TEntity : IAggregateRoot<TId>
     {
-        Task DeleteAsync(TId id, CancellationToken cancellationToken = default);
+        Task<long> CountAsync(CancellationToken cancellationToken = default);
     }
 }
