--- src/Repositories/IEntityAsyncRepositoryBase.cs
@@ -1,10 +1,0 @@
-﻿using PowerUtils.BuildingBlocks.Domain;
-
-namespace PowerUtils.BuildingBlocks.Data.Repositories
-{
-    public interface IEntityAsyncRepositoryBase<TEntity, TId> :
-        IRepositoryBase
-
-        where TEntity : IAggregateRoot<TId>
-    { }
-}
